Debugging react-look, SSR still non-functional
I believe I know why, but I don't know how: react-look depends upon the creation
of a global singleton called 'StyleContainer' that holds all global styles.

One singleton is being created in the createSSR() context and is being created when
the worker starts. The other is being created at render time.

The call to StyleSheet.renderToString() within createSSR() is a accessing the
StyleContainer singleton that does not contain the styles from the render.

It's very frustrating.