New strategy to fix #147; root reducer
See Dan Abramov's answer to http://stackoverflow.com/questions/35622588/how-to-reset-the-state-of-a-redux-store

This fix is much more straightforward than replacing reducers. It simply uses the same
ol' pure function pattern we know and love in redux to return a reset state.

Added whitelisting feature to preserve notficiatoins and unittests for good measure.

Closes #147.