Partially working user impersonation.
We're close on this, but we're not quite there yet.

When we assume the role of the user to impersonate and redirect,
we're missing a few objects from cashay.entities:

   * User
   * Presence

I'm guessing this is because the 'updateUserWithAuthToken' mutation
needs to be paired with the propery query so these fields end up in
the cache.