Don't collapse card for remote spotlight drag and avoid timeout
Regularly send updates that spotlight is still open to keep the remote
end from timing out the drag and closing it. In order to actually update
the timeout a field `updatedAt` was added to the payload which sole
purpose is that something changes and triggers a component rerender.
Ideally this logic would be moved to a lower level where this is not
needed.