Show remote spotlight activity (#5500)
* Show remote spotlight activity

Start moving the card around if the remote end opens spotlight.

Missing:
* keep a placeholder in the list, otherwise the card may obstruct other
  tasks
* send regular updates to avoid timeout on the client side

* Don't collapse card for remote spotlight drag and avoid timeout

Regularly send updates that spotlight is still open to keep the remote
end from timing out the drag and closing it. In order to actually update
the timeout a field `updatedAt` was added to the payload which sole
purpose is that something changes and triggers a component rerender.
Ideally this logic would be moved to a lower level where this is not
needed.

* Make clientHeight and clientWidth optional in updateDragLocation

Make clientHeight and clientWidth optional since these are not required
for spotlight. Instead validate that these are set in case isSpotlight
!== true.

* Cleanup

* Tweak the animation for remote spotlight

* Close spotlight on drag conflict if the remote end won

Moved mutating the store down from GroupingKanban to better resolve the
conflict.

* Cleanup

* Don't send regular updates for open spotlight

This reduces the network traffic with the downside of not detecting a
stale open spotlight, which will block a task for up to 2 minutes if a
participant is disconnected.

* Cancel remote drag if participant disconnected from meeting