Cleanup TaskFooterIntegrateMenu
This menu has only 2 main states now: push with integration or add
integration. Showing a not integrated message was omitted and replaced
with the option to add one. Each menu has an optional label now
displayed to show some context, i.e. whose credentials are used or why
the user would want to add an integration.