Copy integration first before moving a task to a different team
If a task is integrated and the user does not have an integration set up
for the target team, then copy their source teams integration over.

If they have neither source nor target team integration and the task was
pushed by someone else, then check if their integration is present in
target team and use that.