Allow moving integration tasks between teams (#5513)
* Allow moving integration tasks between teams

If a task with the same hash exists in the destination, it is deleted
first.

This currently only touches `changeTaskTeam`, but moving teams is also
possible via `updateTask`.

* Remove teamId from updateTask mutation

This removes the ability of moving tasks between teams via `updateTask`
and requires `changeTaskTeam` to be called. Having only one simpler
mutation to do this is less error prone and reduces maintenance.

* Before moving a task to another team check if an integration is needed

If the task has an integration, then the target team user must have the
right integration set up to ensure the task can be displayed in the
target team.

* Fix TaskFooterTeamAssigneeMenu_task fragment

* Show add integration dialog when moving task to team w/o integration

Before trying to move an integrated task to another team, the client
side checks if the user has set up the correct integration for the
target team. If not a dialog is presented to the user to add it on the
spot.

* Polish add integration now dialog

* Cleanup

* Copy integration first before moving a task to a different team

If a task is integrated and the user does not have an integration set up
for the target team, then copy their source teams integration over.

If they have neither source nor target team integration and the task was
pushed by someone else, then check if their integration is present in
target team and use that.

* Don't start OAuth flow when moving a to another team w/o integration

Assume for now in the UI that a user has set up the integration for the
source team and that we can just transfer it over to the target team.

* Cleanup

* Clear freshly added integration auth from dataLoader

When a task changes team and the integration is copied to the new team,
then we already checked the integration for the target team, thus the
dataLoader has `null` cached for it. We need to clear this value before
publishing/returning any results which might read that value again.

Co-authored-by: Matt Krick <matt.krick@gmail.com>