Make return type optional in primaryKeyDataLoader
If no item exists with the key, undefined is returned. This was already
done for some loaders (e.g. Users).
If the id is an input of the resolver, the call site must check the
return, otherwise we assume database integrety and assert the value.
ForeignKeyDataLoader is not affected since it returns arrays for each
key, which results in just an empty array.