Add primary and foreign key loaders (#5847)
* Fix tests

* Make postgres queries more consistent

* Add pgPrimaryKeyLoaderMaker

The loader maker uses the generated Postgres query directly. The
generated query is used to have the generated type. However the types do
not match the types actually in use, thus an additional type argument is
added to cast it.

* Add foreignKeyLoaderMaker and make pgPrimaryKeyLoaderMaker generic

pgPrimaryKeyLoaderMaker is a bit clunky since the generated postgres
type does not always match the one actually used. Thus it's easier to
make primaryKeyLoaderMaker more generic and pass the wrapped query as
loader function instead. This way it could also be used for non-pg
types.

* Make return type optional in primaryKeyDataLoader

If no item exists with the key, undefined is returned. This was already
done for some loaders (e.g. Users).
If the id is an input of the resolver, the call site must check the
return, otherwise we assume database integrety and assert the value.
ForeignKeyDataLoader is not affected since it returns arrays for each
key, which results in just an empty array.

* Don't notify if team could not be loaded

This should never occur, but in case it does, we shouldn't show 'Team:
undefined' to the customer.

* Add NullableDataloader with convenience loadNonNull function

This can be used for most db loader functions which may return
undefined. On the call site we may have more information like knowing a
foreign key ensures the value existed, thus might want to assert
non-nullness.

* Remove isNeitherNullNorError in favour of isValid