Add JiraServer integration provider fields (#5874)
* Add JiraServer integration provider fields

Add required fields to the database and `addIntegrationProvider`
mutation so the provider can be added.

* Update types/graphql.ts

* Don't assume there is no global OAuth1 provider

* Fix AddIntegrationProviderPayload before it's too late

A new integration provider is available for the whole team or org. It's
not important to add the user on it, since we want to be able to add an
integration for others.

* Rename privateKey field to consumerSecret

This is more standard OAuth1 naming and how the secret is interpreted is
up to the service itself.