feat(poker): Add JiraServer field selection
JiraServerRestManager pulls information about available fields per issue
type (instead of per issue). This is done so the result can be cached by
the dataloader.
User setting is also stored per issue type. This means if there are
multiple issue types with a story point field, the user must set it per
issue type. A possible future improvement could check if the field is
available on other issue types as well and if those don't have a setting
then we could apply this setting there as well.