feat(notifications): Refactor Slack/Mattermost into NotificationHelper (#6262)
* Refactor Slack/Mattermost notifications into NotificationHelper

A common `NotificationHelper` loads the settings and sends the message
via the individual `NotificationIntegrationHelper`s. These then don't
need to load any additional data and are just responsible for sending
out single messages.

* Use same type for IntegrationNotifier, SlackNotifier and MattermostNotifier

All notifiers now use the same signature. These are not classes because
the authentication to load depends partially on the event which is being
sent.

* Add EmptyNotifier to allow type safe iteration over methods

* Simplify Notifier at the expense of repetetion