feat(Jira Server): Voting to different fields in Sprint Poker is now supported (#6437)
* feat(poker): Add JiraServer field selection

JiraServerRestManager pulls information about available fields per issue
type (instead of per issue). This is done so the result can be cached by
the dataloader.
User setting is also stored per issue type. This means if there are
multiple issue types with a story point field, the user must set it per
issue type. A possible future improvement could check if the field is
available on other issue types as well and if those don't have a setting
then we could apply this setting there as well.

* Make implementation match commit message

The Jira Server estimate dimension field is stored per issue type. The
issue type itself is not stored in our database since it is possible to
change it after creation of the issue.

* Cleanup

* use useFragmend instead of createFragmentContainer