fix(Sprint Poker): An exception could occur when modifying the scope in fast succession (#6599)
* Ensure redisQueueLock is always unlocked in updatePokerScope

If there is an exception or an early return in updatePokerScope the lock
wouldn't be unlocked until it timed out. This takes 3s and a new task
tries only for 10s to aquire the lock, so a few of those could time out
a lot of other tasks.

* Work around undefined dimensionRef in UpdatePokerScopeMutation

When updates are done in fast succession the store appears to be in an
undefined state where multiple stages for one taskId are found without
any dimensionRef. As a workaround fall back to the placeholder
dimensionRef in that case.