feat: New meeting dialog is now a modal
This is the first step in redesigning the dialog. The dialog still uses
the same route as before. If the route is hit directly, the meetings
dashboard is rendered in the background. If it's reached by pressing the
"Add meeting" button, then the current page is put into the location
state to be rendered as background.