feat: New meeting type carousel (#6950)
* feat: New meeting dialog is now a modal

This is the first step in redesigning the dialog. The dialog still uses
the same route as before. If the route is hit directly, the meetings
dashboard is rendered in the background. If it's reached by pressing the
"Add meeting" button, then the current page is put into the location
state to be rendered as background.

* Fix NewMeeting dialog portal parents

`parentId` needs to be passed for all nested modals/dialogs so these
don't also close the NewMeeting dialog when they're closed.

* Make top bar start meeting button flat

* Fix markup

* history.replace

* Remove cancel button

* Tweak css

* Tweak image border radius

* Remove source

* Header and footer

* New carousel

* Show last meeting first

* Markup small tweaks

* Fix standups description

* Update async standup bg color

* Change default meeting order

* Always slide to clicked slide

* Reduce active meetings padding

* Fix padding

* Make sure description will always be one line

Co-authored-by: Georg Bremer <github@dschoordsch.de>