fix(Jira): show only available fields in Sprint Poker (#7257)
* Read estimation field from new table

When reading the jira estimation for an issueType fall back to the empty
issue type for all accounts which transitioned and have not yet voted on
these fields.

* Remove add missing Jira field functionality

This isn't working in all cases and confusing from a user perspective.

* Remove TeamIntegrations

The TeamIntegrations weren't used anywhere and the integration is always
member specific and the selected fields are issue specific, thus it
shouldn't be placed in Team.

* Kill remains of addMissingJiraField

* Fall back to other issue types when querying JiraDimensionFieldMap

If there is not a match for the current issue type, fall back to all
other ones, sorted by updatedAt, and check if the field exists on the
current issue as well.

* Disable sorting jiraDimensionField migrations

These migrations have type errors now since the field was removed from
the `Team` type. Also sorting is not necessary anymore because the data
is stored in a normalized way in Postgres.

* Allow "Story point estimate" for simplified projects and vice versa

Also allow "Story Points" for legacy projects. There is nothing stopping
the user from adding those fields in either project type, thus we
shouldn't disallow them.

* Remove ensureJiraDimensionField logic

This is not required anymore in the current implementation.

* Fix issues with changing fields to "As Comment" or "Do not update"

* Cleanup

* Handle jiraDimensionFields conflicts more gracefully

There are duplicates in production which would fail the migration. Just
ignore those and move on.