fix(Jira): fix Sprint Poker voting for projects with duplicate fields (#7607)
* fix(Jira): fix Sprint Poker voting for projects with duplicate fields

Jira allows to have multiple fields with the same name in a project,
even on the same issue. While this does not make sense from a usability
standpoint, this backfired for us, because we were selecting fields
based on their name and could end up selecting fields which belonged to
a different issue type altogether. Now even multiple fields called
"Story Points" on one issue are allowed and work.

* Cleanup

* Cleanup