Log stripe quantity mismatch
Subscription quantity changes are not updated on a draft invoice and
cannot manually be updated either. So although we're getting a callback
for the draft, we can only detect a mismatch, but not directly correct
it.
It would be possible to apply additional line items to correct the
amount due and then filter these out when presenting the invoice, but
for now let's just log the issue and move on to get something working.