feat: remove proration (#7721)
* chore: Remove proration

* Check and adjust the user count before generating an invoice

For upcoming and due invoices, check the organization user count and if
there is a mismatch with Stripe, update it there.

* Cleanup

* Log stripe quantity mismatch

Subscription quantity changes are not updated on a draft invoice and
cannot manually be updated either. So although we're getting a callback
for the draft, we can only detect a mismatch, but not directly correct
it.
It would be possible to apply additional line items to correct the
amount due and then filter these out when presenting the invoice, but
for now let's just log the issue and move on to get something working.

* Prettier

* Cleanup

* Add testing code

* Redis lock updating the stripe quantity

* Add comment for the updateSubscriptionQuantity locking

* Don't throw an error when not being able to aquire a lock

Add a `tryLock` method to RedisLockQueue which returns whether or not a
lock could be aquired. Use this in updateSubscriptionQuantity to avoid
misleading error messages.

* Shutdown redis after tests