Add `canAccess` field to `User` to check if a value should be non-null
This field is used to determine whether or not an invitation request
should be send out. If the value is false, we send a request, otherwise
we will render it. If the `Team` or `Meeting` is `null` although it
should be accessible, we throw an error on the client side.