chore: Throw and catch GraphQL errors on the client (#7932)
* chore: Throw and catch GraphQL errors on the client

Relay does not report graphql errors (like "Cannot return null for
non-nullable field...").

* Move User graphql definitions out of _legacy

* Add `canAccess` field to `User` to check if a value should be non-null

This field is used to determine whether or not an invitation request
should be send out. If the value is false, we send a request, otherwise
we will render it. If the `Team` or `Meeting` is `null` although it
should be accessible, we throw an error on the client side.

* Check if the entity is present in addition to canAccess

In case the information is updated but the `canAccess` information
wasn't properly refreshed, also check if the entity got valid in the
meantime.

* Update canAccess in the store

* Beautify Atmosphere error handle sink