Add team health phase to all retrospective settings
It's much less error prone to add the new phase to all retrospective
settings by default and filtering the paid phases from the settings.
The alternative of adding it only to paid teams is more tricky as it
would need to be added in all upgrade paths and we'd need to keep track
if it was disabled.