fix: Show correct error message on login failure with invitation (#8726)
Before this fix the wrong error would be shown when a user followed an
invitation link but then failed to sign in. Now we're not throwing a
permission exception anymore in `acceptTeamInvitation` if the caller is
not signed in, but instead return an error result. This way we can
handle it on the client side and the return of the login mutation is not
lost.