fix: Check if team member status on accepting an invite twice (#8735)
When an invite was accepted a second time, we did try to return the team
and meeting information regardless of whether or not the user was still
on the team. This was checked further down in the GraphQL tree, but can
result in misleading errors in Sentry. Checking for this particular case
now to see if it still shows up for users down the road which might
cause to a real bug somewhere.