fix: Do not serve cached opaque responses if possible
We don't know whether an opaque response was a success, so if we're
serving the response from cache indefinetely, we might serve broken
versions till the next update. To avoid this, only serve opaque
responses if the network request threw an exception.
Right now, dynamic images are not cached by the service worker this
way, but the browser should do caching, too. Right now it's not possible
to request these with CORS.