fix: Fix duplicate organizations for teams with auto join
If an organization had multiple teams with `autoJoin` enabled, multiple
`OrganizationUser` objects would be created for a new user.

This fixes the creation of duplicate `OrganizationUser` objects but does
not fix the existing duplicates. Ideally we would add a migration for
this, but filtering over all `OrganizationUser` entries is expensive and
only the Parabol org is affected.