fix: Fetch CORS resources from network
If we fetch our own assets we want a CORS response to properly cache it
in the service worker. If the response is served from the browser cache
however, then there is a high chance that the required response headers
are missing and the browser responds with a CORS error instead of
refetching the resource. This happens because the Vary header does not
include headers not present in the original request.
To avoid this error, always fetch from the network for requests we
intend to cache in the SW.