fix: Threaded reply input does not immediately close
When replying a local update is commited to the store. This happened in
the mouse event handler and is propagated to other components before the
event handling of the click finished. This caused the `useClickAway`
handler to fire for the same click event and revert the store update
immediately. There are two ways to fix this, stop propagating the event
or do the store update after the event handler via setImmediate. I chose
the former.