Add start retrospective endpoint
Also we're using no the public graphql interface with a user token
generated from the submitted email from mattermost. So we still trust
mattermost, but don't want to have the rights escalated so the rest of
the permission checks still work.
The queries need to be compiled to work in production.