fix: Update service worker when sources return 404
If the service worker fetches source files from Parabol which in turn
return 404, then the current sources are probably stale and the service
worker should update or unregister itself.

Also fixed an issue with service worker updates on Safari after a page
refresh. The current service worker registration might not be ready
immediately and thus we need to wait a bit before triggering the update.