fix: save edited task content on drag & drop
The state was wrapped in the Draggable component and thus the old state
was kept by react-beatiful-dnd when dragging and the updates were
discarded.
Because react-beatiful-dnd is no longer maintained I opted to make only
a minimal impact change because if we switch libraries we probably have
to do a bigger refactoring.

There is one caveat with the current change: Because I did not want to
pass the dragging state down all the components, we lost the live
updates of the task status marker (the color bar which indicates the
column). I feel this is more consistent anyways as the task is only
moved once dropped, so I did not invest the effort to implement it.