fix: window.__ACTION__ is not defined for MattermostPlugin
The plugin does not set the object. The easiest way to make client
components work anyways is by setting the object to optional. This is
done in the mattermost plugin, so only client components which are
actually used by mattermost are affected. So far this is only the image
selector.
Also fixed the formatting bubble menu to be somewhat presentable.