fix: only billing users can downgrade a plan
The UI was misleading in that it showed the downgrade button and dialog
also for non-billing leaders which resulted in confusion as they aren't
allowed to downgrade, so the mutation failed.

Additionally we tried to promote the user who upgraded initially to a
billing user, but we were grabbing the userId from the server auth
token. I added the userId to the subscription metadata so we can perform
this promotion now successfully.