fix: give new users access to shared pages
If a user signed up, the access was transferred from PageAccessExternal
to PageAccessUser, but not all the way to PageAccess, thus the user
could not access the page.
Also after login we need to invalidate the store as we might have loaded
pages with different permissions before. Ideally this would be on viewer
which is invalidated on signup, but external users don't have a viewer.