fix: valid Team.agendaItems for addTeam and addOrg
When adding a team we previously sent individual subscription events to
the creator via the Team and Organization subscriptions. This was
possible because a viewer would subscribe with their userId to these
channels as well. This means a user could get notifications for orgs and
teams which were not included in their auth token yet.
Sending these types of notifications only via the notifications
subscription mitigates this. This subscription is updated with the new
AuthTokenPayload.