fix: allow meeting queries with 0 own teams
Normally this would be a user input error, but it can happen naturally
in insights when a page is shared with members outside of a team.
We could either manually filter an check on the client, or simply allow
the meeting query with 0 teams and just return an empty result. The
latter seems cleaner and still correct to me.