fix: reload page to work around Service Worker context closed
The error happens when a service worker update is detected, but the site
is not loaded. When Safari then tries to fetch the main <script> it
fails, because it's intercepted by the current service worker. Because
the load of the script fails, the code for registering the new one is
missing, thus the whole load fails.