fix: toReversed is not a function
Very very outdated chrome version 103 causes this error. Because we only
used it once and it was used on a copy, I replaced it with reverse().
There is no good way to put it on the old browser ignore list, as the
error message is a bit too generic, we could accidentally call
toReversed() on a non-array and cause this.