Move table data into separate map
While it has a few advantages to put the data straight on the root
document, the biggest disadvantage is that fields cannot be removed from
it anymore. That would mean a database could grow over time when data is
added and deleted again.
Instead just use a couple of top level fields.